Create a reusable CollapsibleCard component and refactor UI to use it.
- A new generic `CollapsibleCard.js` component has been created in `src/components` to encapsulate the Bootstrap collapse functionality.
- The `Controls`, `Results`, and `Status` components have been refactored to use this new reusable component.

This change eliminates duplicated code, improves the modularity of the UI, and makes the code easier to maintain, in line with the user's goal of creating more generic 'parts'.
